Updated CSS for more styling flexibility.
I updated the "styles.css" with CSS custom properties to make it easier to change values like the default size of the iframe, etc. There is also a new and modern loading animation and a bunch of new button types, which are Base64 encoded in the CSS. IMPORTANT: Choose the icon style you like and delete the others, because they make the file significantly larger. Additionally, the default max-size for the holder div now increases as the browser size decreases, to use the space more efficiently. This is done with a combination of "calc()" and "clamp()" so it doesn't exceed 100%. Of course, the values are also easily adjustable.